+++ penna-core/src/main/java/penna/core/slf4j/mdc/Mdc.java
@@ -1,176 +1,60 @@
 package penna.core.slf4j.mdc;
 
-import org.slf4j.spi.MDCAdapter;
-import penna.core.internals.MapWrapperRingBuffer;
-import penna.core.internals.MapWrapperTicket;
+import penna.core.internals.store.StringMap;
+import penna.core.internals.store.StringTreeMap;
 
-import java.util.Deque;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.function.BiConsumer;
 
-public sealed interface Mdc extends MDCAdapter {
+/**
+ * This is the MDC Proxy between the adapter and the actual storage.
+ * The purpose of this intermediary layer is to ensure a few things:
+ * <br/>
+ * - We don't create unnecessary garbage in the memory if no values are associated with the MDC, so we reuse the
+ * {@link EmptyMdc} instance in all threads by default;
+ * - We can swap the internal implementation without interfering with the adapter (given it's an external interface
+ * from slf4j);
+ * <br />
+ * If we want to make any configuration changes, the inner {@link Control} class is the one that should be updated
+ */
+public sealed interface Mdc extends Map<String, String> permits EmptyMdc, MdcStorage {
 
     boolean isNotEmpty();
-    int size();
-    void forEach(BiConsumer<String, String> action);
 
-    final class Inner {
-        private Inner() {}
-        private static final EmptyMdc empty = new EmptyMdc();
+    void remove(String key);
+
+    String get(String key);
+
+    default void setContextMap(Map<String, String> contextMap) {
+        if (contextMap instanceof Mdc mdcCtx) {
+            Control.mdcStorage.set(mdcCtx);
+        } else if (contextMap.isEmpty()) {
+            Control.mdcStorage.set(Control.empty);
+        } else if (contextMap instanceof StringMap arrayMap) {
+            Control.mdcStorage.set(new MdcStorage(arrayMap));
+        } else {
+            Control.mdcStorage.set(new MdcStorage(new StringTreeMap(contextMap)));
+        }
+    }
+
+
+    Map<String, String> getCopyOfContextMap();
+
+    /**
+     * Basic static storage to allow for control over the MDC structure;
+     */
+    final class Control {
+        private Control() {}
+
+        /**
+         * {@link EmptyMdc} Singleton instance. Default visibility so the {@link Mdc} implementations
+         * can refer to it.
+         */
+        static final EmptyMdc empty = new EmptyMdc();
+
+        /**
+         * The actual MDC storage for current thread
+         */
         public static final ThreadLocal<Mdc> mdcStorage = ThreadLocal.withInitial(() -> empty);
-
     }
 
-    final class EmptyMdc implements Mdc {
-        private static final Map<String, String> empty = Map.of();
-
-        @Override
-        public void put(String key, String val) {
-            Inner.mdcStorage.set(new MdcStorage(key, val));
-        }
-
-        @Override
-        public String get(String key) {
-            return null;
-        }
-
-        @Override
-        public void remove(String key) {}
-
-        @Override
-        public void clear() {}
-
-        @Override
-        public Map<String, String> getCopyOfContextMap() {
-            return empty;
-        }
-
-        @Override
-        public void setContextMap(Map<String, String> contextMap) {
-            Inner.mdcStorage.set(new MdcStorage(contextMap));
-        }
-
-        @Override
-        public void pushByKey(String key, String value) {
-            // Swap
-        }
-
-        @Override
-        public String popByKey(String key) {
-            return null;
-        }
-
-        @Override
-        public Deque<String> getCopyOfDequeByKey(String key) {
-            return null;
-        }
-
-        @Override
-        public void clearDequeByKey(String key) {}
-
-        @Override
-        public boolean isNotEmpty() {
-            return false;
-        }
-
-        @Override
-        public int size() {
-            return 0;
-        }
-
-        @Override
-        public void forEach(BiConsumer<String, String> action) {}
-    }
-    final class MdcStorage implements Mdc {
-        final MapWrapperRingBuffer ringBuffer = new MapWrapperRingBuffer();
-
-        MdcStorage(String key, String val) {
-            ringBuffer.put(new HashMap<>());
-            ringBuffer.get().inner().put(key, val);
-        }
-
-        MdcStorage(Map<String, String> context) {
-            ringBuffer.put(context);
-        }
-
-        @Override
-        public void put(String key, String val) {
-            ringBuffer.get().inner().put(key, val);
-        }
-
-        @Override
-        public String get(String key) {
-            return ringBuffer.get().get(key);
-        }
-
-        @Override
-        public void remove(String key) {
-            ringBuffer.get().inner().remove(key);
-        }
-
-        @Override
-        public void clear() {
-            Inner.mdcStorage.set(Inner.empty);
-        }
-
-        @Override
-        public Map<String, String> getCopyOfContextMap() {
-            var wrapper = ringBuffer.get();
-            ringBuffer.put(new HashMap<>(wrapper.inner()));
-
-            return wrapper;
-        }
-
-        @Override
-        public void setContextMap(Map<String, String> contextMap) {
-            if (contextMap instanceof MapWrapperTicket wrapper) {
-                if (ringBuffer.get(wrapper.ticket()) == wrapper) {
-                    ringBuffer.reset(wrapper.ticket());
-                } else {
-                    ringBuffer.put(wrapper.inner());
-                }
-            } else {
-                ringBuffer.put(contextMap);
-            }
-        }
-
-        @Override
-        public void pushByKey(String key, String value) {
-
-        }
-
-        @Override
-        public String popByKey(String key) {
-            return null;
-        }
-
-        @Override
-        public Deque<String> getCopyOfDequeByKey(String key) {
-            return null;
-        }
-
-        @Override
-        public void clearDequeByKey(String key) {
-
-        }
-
-        @Override
-        public boolean isNotEmpty() {
-            return !ringBuffer.get().isEmpty();
-        }
-
-        @Override
-        public int size() {
-            return ringBuffer.get().size();
-        }
-
-        @Override
-        public void forEach(BiConsumer<String, String> action) {
-            var map = ringBuffer.get();
-            for (var key : map.keySet()) {
-                action.accept(key, map.get(key));
-            }
-        }
-    }
 }