+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) with an addendum:
 the PATCH component is omitted when its value is `0`.
 
+## Unreleased
+
+### `penna-core`
+
+#### Changed
+
+- Replace internal MDC storage implementation([#83](https://github.com/hkupty/penna/pull/83))
+
 ## 0.7.2
 
 ### `penna-core`